+++ Libplanet.Explorer.Executable/Program.cs
@@ -283,11 +283,13 @@ If omitted (default) explorer only the local blockchain store.")]
                             : default(AppProtocolVersion),
                     };
 
+                    var hostOptions = new HostOptions(null, new[] { options.IceServer });
+
                     swarm = new Swarm<NullAction>(
                         blockChain,
                         privateKey,
                         apvOptions,
-                        iceServers: new[] { options.IceServer },
+                        hostOptions,
                         options: swarmOptions
                     );
                 }