+++ Libplanet.Net.Tests/SwarmTest.cs
@@ -289,9 +289,13 @@ namespace Libplanet.Net.Tests
         public async Task MaintainStaticPeers()
         {
             var keyA = new PrivateKey();
+            var hostOptionsA = new HostOptions(
+                IPAddress.Loopback.ToString(), new IceServer[] { }, 20_000);
+            var hostOptionsB = new HostOptions(
+                IPAddress.Loopback.ToString(), new IceServer[] { }, 20_001);
 
-            Swarm<DumbAction> swarmA = CreateSwarm(keyA, listenPort: 20000);
-            Swarm<DumbAction> swarmB = CreateSwarm(listenPort: 20001);
+            Swarm<DumbAction> swarmA = CreateSwarm(keyA, hostOptions: hostOptionsA);
+            Swarm<DumbAction> swarmB = CreateSwarm(hostOptions: hostOptionsB);
             await StartAsync(swarmA);
             await StartAsync(swarmB);
 
@@ -330,7 +334,7 @@ namespace Libplanet.Net.Tests
             Assert.DoesNotContain(swarmA.AsPeer, swarm.Peers);
             Assert.Contains(swarmB.AsPeer, swarm.Peers);
 
-            Swarm<DumbAction> swarmC = CreateSwarm(keyA, listenPort: 20000);
+            Swarm<DumbAction> swarmC = CreateSwarm(keyA, hostOptions: hostOptionsA);
             await StartAsync(swarmC);
             await AssertThatEventually(() => swarm.Peers.Contains(swarmB.AsPeer), 5_000);
             await AssertThatEventually(() => swarm.Peers.Contains(swarmC.AsPeer), 5_000);
@@ -533,30 +537,21 @@ namespace Libplanet.Net.Tests
             var key = new PrivateKey();
             var apv = AppProtocolVersion.Sign(key, 1);
             var apvOptions = new AppProtocolVersionOptions() { AppProtocolVersion = apv };
+            var hostOptions = new HostOptions(
+                IPAddress.Loopback.ToString(), new IceServer[] { });
 
             Assert.Throws<ArgumentNullException>(() =>
-                new Swarm<DumbAction>(null, key, apvOptions));
+                new Swarm<DumbAction>(null, key, apvOptions, hostOptions));
             Assert.Throws<ArgumentNullException>(() =>
-                new Swarm<DumbAction>(blockchain, null, apvOptions));
-
-            // Swarm<DumbAction> needs host or iceServers.
-            Assert.Throws<ArgumentException>(() =>
-                new Swarm<DumbAction>(blockchain, key, apvOptions));
-
-            // Swarm<DumbAction> needs host or iceServers.
-            Assert.Throws<ArgumentException>(() =>
-                new Swarm<DumbAction>(
-                    blockchain,
-                    key,
-                    apvOptions,
-                    iceServers: new IceServer[] { }));
+                new Swarm<DumbAction>(blockchain, null, apvOptions, hostOptions));
         }
 
         [Fact(Timeout = Timeout)]
         public void CanResolveEndPoint()
         {
             var expected = new DnsEndPoint("1.2.3.4", 5678);
-            using (Swarm<DumbAction> s = CreateSwarm(host: "1.2.3.4", listenPort: 5678))
+            var hostOptions = new HostOptions("1.2.3.4", new IceServer[] { }, 5678);
+            using (Swarm<DumbAction> s = CreateSwarm(hostOptions: hostOptions))
             {
                 Assert.Equal(expected, s.EndPoint);
                 Assert.Equal(expected, s.AsPeer?.EndPoint);
@@ -597,9 +592,11 @@ namespace Libplanet.Net.Tests
         public async Task ExchangeWithIceServer()
         {
             var iceServers = FactOnlyTurnAvailableAttribute.GetIceServers();
-            var seed = CreateSwarm(host: "localhost");
-            var swarmA = CreateSwarm(iceServers: iceServers);
-            var swarmB = CreateSwarm(iceServers: iceServers);
+            var seedHostOptions = new HostOptions("localhost", ImmutableList<IceServer>.Empty, 0);
+            var swarmHostOptions = new HostOptions(null, iceServers);
+            var seed = CreateSwarm(hostOptions: seedHostOptions);
+            var swarmA = CreateSwarm(hostOptions: swarmHostOptions);
+            var swarmB = CreateSwarm(hostOptions: swarmHostOptions);
 
             try
             {
@@ -652,18 +649,16 @@ namespace Libplanet.Net.Tests
             string username = userInfo[0];
             string password = userInfo[1];
             var proxyUri = new Uri($"turn://{username}:{password}@localhost:{port}/");
-
-            IEnumerable<IceServer> iceServers = new[]
-            {
-                new IceServer(url: proxyUri),
-            };
+            IEnumerable<IceServer> iceServers = new[] { new IceServer(url: proxyUri) };
 
             var cts = new CancellationTokenSource();
             var proxyTask = TurnProxy(port, turnUrl, cts.Token);
 
             var seedKey = new PrivateKey();
-            var seed = CreateSwarm(seedKey, host: "localhost");
-            var swarmA = CreateSwarm(iceServers: iceServers);
+            var seedHostOptions = new HostOptions("localhost", ImmutableList<IceServer>.Empty, 0);
+            var swarmHostOptions = new HostOptions(null, iceServers, 0);
+            var seed = CreateSwarm(seedKey, hostOptions: seedHostOptions);
+            var swarmA = CreateSwarm(hostOptions: swarmHostOptions);
 
             async Task RefreshTableAsync(CancellationToken cancellationToken)
             {