+++ Libplanet.Net.Tests/Transports/BoundPeerExtensionsTest.cs
@@ -1,4 +1,3 @@
-#nullable disable
 using System;
 using System.Net;
 using System.Net.Sockets;
@@ -29,13 +28,13 @@ namespace Libplanet.Net.Tests.Transports
             var fx = new MemoryStoreFixture();
             var policy = new BlockPolicy<DumbAction>();
             var blockchain = MakeBlockChain(policy, fx.Store, fx.StateStore);
-            var apvKey = new PrivateKey();
             var swarmKey = new PrivateKey();
-            var apv = AppProtocolVersion.Sign(apvKey, 1);
+            var apv = AppProtocolVersion.Sign(new PrivateKey(), 1);
             var apvOptions = new AppProtocolVersionOptions() { AppProtocolVersion = apv };
-
             string host = IPAddress.Loopback.ToString();
             int port = FreeTcpPort();
+            var hostOptions = new HostOptions(
+                IPAddress.Loopback.ToString(), new IceServer[] { }, port);
 
             var option = new SwarmOptions();
 
@@ -43,8 +42,7 @@ namespace Libplanet.Net.Tests.Transports
                 blockchain,
                 swarmKey,
                 apvOptions,
-                host: host,
-                listenPort: port,
+                hostOptions,
                 options: option))
             {
                 var peer = new BoundPeer(swarmKey.PublicKey, new DnsEndPoint(host, port));